+++ integrations/gitlab/gitlab_integration/events/hooks/jobs.py
@@ -2,7 +2,7 @@ from typing import Any
 
 from gitlab.v4.objects import Project
 
-from gitlab_integration.events.hooks.project import ProjectHandler
+from gitlab_integration.events.hooks.base import ProjectHandler
 from gitlab_integration.utils import ObjectKind
 from port_ocean.context.ocean import ocean
